chore: align crate names with kebab-case directories and fix workspace errors

+++ 189-rotate-array/src/lib.rs
@@ -2,8 +2,8 @@ pub struct Solution;
 
 impl Solution {
     pub fn rotate(nums: &mut Vec<i32>, k: i32) {
-        let n = nums.len() as i32;
-        let k = k % n;
+        let n = nums.len();
+        let k = (k as usize) % n;
 
         if k == 0 {
             return;
